+++ public_html/index.php
@@ -39,10 +39,6 @@ if (!in_array('sitemap', $_PLUGINS) || !SMAP_canView()) {
     exit;
 }
 
-// Loads config
-USES_sitemap_class_config();
-smapConfig::loadConfigs();
-
 //===========================
 //  Functions
 //===========================
@@ -52,19 +48,18 @@ smapConfig::loadConfigs();
 */
 function SITEMAP_getSelectForm($selected = 'all')
 {
-    global $_CONF, $_SMAP_CONF, $LANG_SMAP, $_SMAP_MAPS, $_SMAP_DRIVERS;
+    global $_CONF, $_SMAP_CONF, $LANG_SMAP;
 
     $this_script = $_CONF['site_url'] . '/sitemap/index.php';
-    $num_drivers = count($_SMAP_MAPS);
+    $drivers = Sitemap\Config::getDrivers();
     $LT = new Template($_CONF['path'] . '/plugins/' . $_SMAP_CONF['pi_name'] . '/templates');
     $LT->set_file('selector', 'selector.thtml');
     $LT->set_var(array(
         'action_url'    => $this_script,
         'all_sel'   => $selected == 'all' ? 'selected="selected"' : '',
-        'num_drivers' => $num_drivers,
     ) );
     $LT->set_block('selector', 'selectOpts', 'opts');
-    foreach ($_SMAP_DRIVERS as $driver) {
+    foreach ($drivers as $driver) {
         $LT->set_var(array(
             'driver_name'   => $driver->getName(),
             'driver_display_name' => $driver->getDisplayName(),
@@ -102,17 +97,11 @@ function SITEMAP_buildItems($driver, $pid)
     } else {
         $sp_except = array();
     }
-    if (version_compare(GVERSION, '2.0.0', '>=')) {
-        $c = \glFusion\Cache::getInstance();
-        $key = $c->createKey('sitemap_' . $driver->getName() . '_' . $pid);
-        if ($c->has($key)) {
-            $items = $c->get($key);
-        } else {
-            $items = $driver->getItems($pid);
-            $c->set($key, $items, array('sitemap', 'sitemap_' . $driver->getName()));
-        }
-    } else {
+    $key = $driver->getName() . '_' . $pid;
+    $items = Sitemap\Cache::get($key);
+    if ($items === NULL) {
         $items = $driver->getItems($pid);
+        Sitemap\Cache::set($key, $items, $driver->getName());
     }
     $num_items = count($items);
     if ($num_items > 0 && is_array($items)) {
@@ -152,17 +141,11 @@ function SITEMAP_buildCategory($driver, $cat)
     $temp = $T->get_var('child_categories');    // Push $T->var('child_categories')
 
     // Builds {child_categories}
-    if (version_compare(GVERSION, '2.0.0', '>=')) {
-        $key = 'sitemap_' . $driver->getName() . '_category_' . $cat['id'];
-        $c = \glFusion\Cache::getInstance();
-        if ($c->has($key)) {
-            $child_categories = $c->get($key);
-        } else {
-            $child_categories = $driver->getChildCategories($cat['id']);
-            $c->set($key, $child_categories, array('sitemap', 'sitemap_' . $driver->getName() . '_category'));
-        }
-    } else {
+    $key = $driver->getName() . '_category_' . $cat['id'];
+    $child_categories = Sitemap\Cache::get($key);
+    if ($child_categories === NULL) {
         $child_categories = $driver->getChildCategories($cat['id']);
+        Sitemap\Cache::set($key, $child_categories, $driver->getName());
     }
     if (count($child_categories) > 0) {
         $child_cats = '';
@@ -180,7 +163,6 @@ function SITEMAP_buildCategory($driver, $cat)
             'child_categories', $child_cats
         );
     }
-
     // Builds {category}
     if ($cat['title'] == '') {
         // If an empty category title comes in, default to 'Uncategorized'
@@ -214,9 +196,6 @@ function SITEMAP_buildCategory($driver, $cat)
 //  Main
 //=====================================
 
-// Make the sitemap base class available
-USES_sitemap_class_base();
-
 // Retrieves vars
 $selected = 'all';
 if (isset($_GET['type'])) {
@@ -236,25 +215,14 @@ $T->set_file(array(
 // Load up an array containing all the enabled sitemap classes.
 // Used below to write the sitemap and in the selection creation above.
 // Ensures that only valid driver classfiles are used.
-global $_SMAP_DRIVERS;
-$_SMAP_DRIVERS = array();
-foreach ($_SMAP_MAPS as $pi_name=>$pi_config) {
-    if ($pi_config['html_enabled'] == 0) continue;
-    $classfile = smapConfig::getClassPath($pi_name);
-    if ($classfile) {
-        include_once $classfile;
-        $cls = 'sitemap_' . $pi_name;
-        if (class_exists($cls)) {
-            $_SMAP_DRIVERS[] = new $cls();
-        }
-    }
-}
 
-foreach ($_SMAP_DRIVERS as $driver) {
+$drivers = Sitemap\Config::getDrivers();
+foreach ($drivers as $driver) {
     $num_items = 0;
 
-    // Only display selected driver, or "all"
-    if ($selected != 'all' && $selected != $driver->getName()) {
+    // Only display enabled selected driver, or "all"
+    if ( !$driver->html_enabled ||
+        ($selected != 'all' && $selected != $driver->getName()) ) {
         continue;
     }